% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.example-package}
\alias{blockr.example}
\alias{blockr.example-package}
\title{blockr.example: Example Data Blocks for blockr}
\description{
A minimal blockr package demonstrating essential patterns. Provides a simple data block that serves as a template for creating new blockr packages.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com}

}
\keyword{internal}
