% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{new_example_head_block}
\alias{new_example_head_block}
\title{Head block}
\usage{
new_example_head_block(n = 6L, ...)
}
\arguments{
\item{n}{Number of rows}
}
\description{
An example head block to demonstrate the absolute basics of creating your
own block.
}
