% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.example-package}
\alias{blockr.example}
\alias{blockr.example-package}
\title{blockr.example: An Example Blockr Package With Dummy Blocks And Tests}
\description{
An example blockr package with dummy blocks and tests.
}
\author{
\strong{Maintainer}: Mike PAge \email{mike@cynkra.com}

}
\keyword{internal}
